fix: records reader IT in own file
