+++ tasklist/qa/integration-tests/src/test/java/io/camunda/tasklist/zeebeimport/opensearch/RecordsReaderOpensearchIT.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.tasklist.zeebeimport.opensearch;
+
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
+
+import io.camunda.tasklist.qa.util.TestUtil;
+import io.camunda.tasklist.util.TasklistZeebeIntegrationTest;
+import io.camunda.tasklist.zeebeimport.RecordsReaderHolder;
+import io.camunda.webapps.schema.descriptors.tasklist.index.TasklistImportPositionIndex;
+import io.camunda.webapps.schema.entities.operate.ImportPositionEntity;
+import java.time.Duration;
+import org.awaitility.Awaitility;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.opensearch.client.opensearch.OpenSearchClient;
+import org.opensearch.client.opensearch.core.SearchRequest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+
+public class RecordsReaderOpensearchIT extends TasklistZeebeIntegrationTest {
+
+  @Autowired
+  @Qualifier("tasklistOsClient")
+  OpenSearchClient openSearchClient;
+
+  @Autowired private RecordsReaderHolder recordsReaderHolder;
+  @Autowired private TasklistImportPositionIndex importPositionIndex;
+
+  @BeforeAll
+  public static void beforeClass() {
+    assumeTrue(TestUtil.isOpenSearch());
+  }
+
+  @Test
+  public void shouldWriteDefaultEmptyDefaultImportPositionDocumentsOnRecordReaderStart() {
+    Awaitility.await()
+        .atMost(Duration.ofSeconds(60))
+        .until(
+            () -> {
+              final var searchRequest =
+                  new SearchRequest.Builder()
+                      .size(100)
+                      .index(importPositionIndex.getFullQualifiedName())
+                      .build();
+              final var documents =
+                  openSearchClient.search(searchRequest, ImportPositionEntity.class);
+
+              // all initial import position documents created for each record reader
+              return documents.hits().hits().size()
+                  == recordsReaderHolder.getAllRecordsReaders().size();
+            });
+  }
+}
